[TEST] header의 알림아이콘 로직 테스트중
