+++ src/types/flirtingListType.ts
@@ -30,13 +30,13 @@ export type MyType = {
 };
 
 export type FlirtingListInNotificationType = {
-  created_at: Date | null;
-  flirting_message: string | null;
-  id: number | null;
-  status: string | null;
-  sender_is_read_in_noti: boolean | null;
-  receiver_is_read_in_noti: boolean | null;
-  receiver_uid: string | null;
-  sender_uid: string | null;
+  created_at: Date;
+  flirting_message: string;
+  id: number;
+  status: string;
+  sender_is_read_in_noti: boolean;
+  receiver_is_read_in_noti: boolean;
+  receiver_uid: string;
+  sender_uid: string;
   custom_users: MyType;
 };
